# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Token Count Service API",
    "version": "0.1.0",
    "description": "A free service that offers API endpoints for users to send documents and have them broken down into a token count analysis."
  },
  "servers": [
    {
      "url": "http://localhost:4001",
      "description": "Development server"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Health check endpoint",
        "description": "Returns the status of the Token Count Service with API documentation and endpoint information",
        "operationId": "healthCheck",
        "responses": {
          "200": {
            "description": "Service is running",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Token Count Service is running"
                    },
                    "documentation": {
                      "type": "string",
                      "example": "http://localhost:4001/docs"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database connection failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Failed to connect to the database"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/analyze": {
      "post": {
        "summary": "Analyze a document",
        "description": "Analyzes a document and returns the token count and other analysis.",
        "operationId": "analyzeDocument",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document analyzed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Analysis"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/status": {
      "get": {
        "summary": "Get document analysis status",
        "description": "Retrieve the current status of a document analysis.",
        "operationId": "getDocumentStatus",
        "parameters": [
          {
            "name": "documentId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the document to check."
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the user who owns the document."
          }
        ],
        "responses": {
          "200": {
            "description": "Document status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Analysis"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Document not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
      "Analysis": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "userId": {
            "type": "string"
          },
          "tokenCount": {
            "type": "integer"
          },
          "analysis": {
            "type": "object",
            "properties": {
              "tokenCount": {
                "type": "integer"
              },
              "byteSize": {
                "type": "integer"
              },
              "charCount": {
                "type": "integer"
              },
              "wordCount": {
                "type": "integer"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}